+++ src/firebase-app.ts
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-import {
-  Credential, GoogleOAuthAccessToken, getApplicationDefault
-} from './credential/credential';
+import { Credential, GoogleOAuthAccessToken } from './credential/credential';
+import { getApplicationDefault } from './credential/credential-internal';
 import * as validator from './utils/validator';
 import { deepCopy, deepExtend } from './utils/deep-copy';
 import { FirebaseServiceInterface } from './firebase-service';
